powerpc: plumbing to track the allocation status of all pkeys
Initial plumbing to manage all the keys supported by the
hardware.

Total 32 keys are supported on powerpc. However pkey 0,1
and 31 are reserved. So effectively we have 29 pkeys.

This patch keeps track of reserved keys, allocated  keys
and keys that are currently free.

Also it  adds  skeletal  functions  and macros, that the
architecture-independent code expects to be available.

Signed-off-by: Ram Pai <linuxram@us.ibm.com>

Conflicts:
	arch/powerpc/include/asm/mmu_context.h
	arch/powerpc/include/asm/pkeys.h