powerpc: enable pkeys consulting device tree on LPAR
PAPR defines 'ibm,processor-storage-keys' property. It exposes
two value. The first value indicates the number of data-access
keys  and  the  second    one    indictes    the   number   of
instruction-access keys. This indicates that keys can be data
access only or instructon access only, a indication which is
conflicting with the way PowerISA3.0 defines keys.
This patch does not depend on those numbers, but uses them as
a indication that protection-keys are supported by the MMU.

Non PAPR platforms do not define this property in the device
tree.

Signed-off-by: Ram Pai <linuxram@us.ibm.com>