powerpc: shifted-by-one hidx value
0xf is considered invalid hidx value. It indicates absence of a backing
HPTE for that slot. Everytime a PTE is demoted to 4k backing HPTE, the
secondary-part of the PTE is initialized to all ones.  This patch shifts
the representation by one-modulo-0xf.  hidx 0 gets respresented as 1, 1
as 2,... , and 0xf as 0.  This convention lets us initialize the
secondary-part of the PTE to all zeroes. PTEs are anyway zero'd when
allocated. We dont have to zero them gain; thus have the potential to
save on the initialization. That goes the theory.

However...all theories do not monetize in reality. We already save
initialization if the PTE is freshly created to back 4k-HPTE.  The
saving potential is only for those PTEs that get demoted from 64k-backed
HPTE to 4k-backed HPTE. Unfortunately if the PTE was backed by 64k-HPTE
the secondary-part of the PTE already would have a few bits set to
indicate the hidx of that HPTE. It needs to be cleared.
demoted. This means we have to initialize it and hence NO savings.

Sending this patch as proof-of-concept. I strongly feel its not worth
the complexity for no apparent gain.

Signed-of-by: Ram Pai <linuxram@us.ibm.com>