selftests/vm: fix number of powerpc PKEYS and pkey_assert() condition
The number of reserved PKEYS in PowerNV is different from PowerVM/KVM.
Moreover, this patch also makes it clear that the -1 in the comparison
condition with the variable i on test_pkey_alloc_exhaust() is due to the
range [0;NR_PKEYS-1] and not due to an extra exec-only key.

Tested on powervm and powernv.

Fixes: a3e2d85461c3 ("selftests/vm: fix an assertion in test_pkey_alloc_exhaust()")
Signed-off-by: "Desnes A. Nunes do Rosario" <desnesn@linux.vnet.ibm.com>
Signed-off-by: Ram Pai <linuxram@us.ibm.com>